Created a function to generate JWTs
